icons are same width, moved settings for mobile
