+++ components/home/NavBar.tsx
@@ -2,18 +2,13 @@ import { useSideBarOptionSetter } from '@/context/SideBarOptionCtx';
 import FriendsIcon from '@/components/icons/FriendsIcon';
 import ServersIcon from '@/components/icons/ServersIcon';
 import { useState } from 'react';
-import { useSetChannel, useUserRef } from '@/lib/store';
-import GearIcon from '../icons/GearIcon';
-import EditUserModal from './EditUserModal';
+import { useSetChannel } from '@/lib/store';
 
 export default function NavBar({ type }: { type: 'vertical' | 'bottom' }) {
   const setSideBarOption = useSideBarOptionSetter();
 
   const setChannel = useSetChannel();
-  const editUser = useUserRef();
 
-
-  const [showEditUser, setShowEditUser] = useState(false);
   const [friendsHover, setFriendsHover] = useState(false);
   const [serversHover, setServersHover] = useState(false);
   const bottomStyles = 'hover:cursor-pointer flex justify-center items-center';
@@ -70,22 +65,6 @@ export default function NavBar({ type }: { type: 'vertical' | 'bottom' }) {
       >
         <MessagesIcon hovered={messagesHover} width={6} height={6} />
       </div> */}
-      <div className={`${
-        type == 'bottom' && `${bottomStyles} col-start-8 col-end-10`}`}>
-        {type == 'bottom' &&
-      <>
-        <EditUserModal 
-          showModal={showEditUser} 
-          setShowModal={setShowEditUser} 
-          user={editUser}/>
-        <button 
-          className="w-7 h-7 hover:text-grey-400"
-          onClick={() => {setShowEditUser(true);}}>
-          <GearIcon width={6} height={6}/>
-        </button>
-      </>
-        }
-      </div>
     </>
   );
 }
