+++ components/icons/ChannelMediaIcon.tsx
@@ -4,7 +4,7 @@ export function ChannelMediaIcon({ size = 4 }: { size?: number }) {
       xmlns="http://www.w3.org/2000/svg"
       fill="none"
       viewBox="0 0 256 256"
-      strokeWidth={2}
+      strokeWidth={1}
       stroke="currentColor"
       className={`w-${size} h-${size}`}
     >
