+++ components/icons/ChannelMessageIcon.tsx
@@ -4,7 +4,7 @@ export default function ChannelMessageIcon({ size = '4' }: { size?: string }) {
       xmlns="http://www.w3.org/2000/svg"
       fill="none"
       viewBox="0 0 24 24"
-      strokeWidth={2}
+      strokeWidth={1}
       stroke="currentColor"
       className={`w-${size} h-${size}`}
     >
