+++ components/icons/CheckIcon.tsx
@@ -6,6 +6,7 @@ export function CheckIcon({ className }: { className?: string }) {
       height="24"
       viewBox="0 0 24 24"
       className={className}
+      strokeWidth={1}
     >
       <path
         d="m10 13.6l5.9-5.9q.275-.275.7-.275t.7.275q.275.275.275.7t-.275.7l-6.6 6.6q-.3.3-.7.3t-.7-.3l-2.6-2.6q-.275-.275-.275-.7t.275-.7q.275-.275.7-.275t.7.275l1.9 1.9Z"
