+++ components/icons/DefaultUserIcon.tsx
@@ -5,7 +5,7 @@ export function DefaultUserIcon({ className }: { className?: string }) {
         xmlns='http://www.w3.org/2000/svg'
         fill='none'
         viewBox='0 0 24 24'
-        strokeWidth={1.5}
+        strokeWidth={1}
         stroke='currentColor'
         className={`w-7 p-2 ${className}`}
       >
