+++ components/icons/EmailIcon.tsx
@@ -5,6 +5,7 @@ export default function EmailIcon() {
     viewBox="0 0 34 28"
     fill="none"
     xmlns="http://www.w3.org/2000/svg"
+    strokeWidth={1}
   >
     <path
       d="M30.3333 0.666626H3.66667C1.83334 0.666626 0.350003 2.16663 0.350003 3.99996L0.333336 24C0.333336 25.8333 1.83334 27.3333 3.66667 27.3333H30.3333C32.1667 27.3333 33.6667 25.8333 33.6667 24V3.99996C33.6667 2.16663 32.1667 0.666626 30.3333 0.666626ZM30.3333 7.33329L17 15.6666L3.66667 7.33329V3.99996L17 12.3333L30.3333 3.99996V7.33329Z"
