+++ components/icons/TrashIcon.tsx
@@ -4,7 +4,7 @@ export default function TrashIcon({ styles = '' }) {
       xmlns="http://www.w3.org/2000/svg"
       fill="none"
       viewBox="0 0 24 24"
-      strokeWidth={1.3}
+      strokeWidth={1}
       stroke="currentColor"
       className={`w-5 h-5 ${styles}`}
     >
