+++ components/icons/VerticalSettingsIcon.tsx
@@ -8,7 +8,7 @@ export default function VerticalSettingsIcon({
       xmlns="http://www.w3.org/2000/svg"
       fill="none"
       viewBox="0 0 24 24"
-      strokeWidth={2}
+      strokeWidth={1}
       stroke={`${hovered ? '#46575e' : 'currentColor'}`}
       className="w-5 h-5"
     >
