+++ components/icons/XIcon.tsx
@@ -6,6 +6,7 @@ export function XIcon({ className }: { className?: string }) {
       height="24"
       viewBox="0 0 24 24"
       className={className}
+      strokeWidth={1}
     >
       <path
         // fill="currentColor"
